+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.14.0
+
+> This release has breaking changes!
+- [BREAKING] moment.format("YY") now **no longer throws exception** when year doesn't fall in range (1970, 2030). Instead, it returns full year. e.g., `DateTime(1969).format("YY"); // 1969`
+- [BREAKING] moment.format("gg") now **no longer throws exception** when ISO week year doesn't fall in range (1970, 2030). Instead, it returns full year. e.g., `DateTime(1969).format("gg"); // 1969`
+- [BREAKING] `DateTime.format(...)` now expect named arguments instead of optional positional arguments
+
 ## 0.13.0
 
 > This release has breaking changes!